Turn on all channels when loading a new image from within the GUI
The selected channels have remained set to the last image when loading an image through the GUI. To highlight the number of channels in the newly loaded image, select them all upon loading while continuing to honor the channel command line setting if loading the image in the GUI as a command-line argument. Also, fix the list of selected channels and the displayed channels in the profiles panel channel dropdown to not include channels selected in the previous image but not available in the current image.