Limit metadata update exception when unable to load the main image
The metadata update exception is supposed to be triggered when the metadata is loaded and flagged for update, but the main image could not be loaded but has also been triggered even when it could not be loaded because the default version number fell within range for the exception. Ensure that the metadata version number is above this default value to avoid this exception, instead returning None when both the metadata and main image could not be loaded. Also fix a missing word in the exception message and change it to a `FileNotFoundError` for more specificity since the exception results from the inability to find the main image file.