Fix inconsistency in imported resolutions nested lists
Store resolutions in the transient import metadata dictionary as a single rather than a nested list since the imported image is assumed to contain only a single series. This change fixes an extra nesting of lists when importing multipage files through the CLI and the inconsistency in list nesting in the input and output dictionary in GUI imports. Also fix the extra nesting when importing directories of images in the CLI based on the current `config.resolutions`.