Initiate import setup when image files cannot be loaded through the GUI
The GUI will attempt to import files when the corresponding image cannot cannot be found during loading. This step can take a considerable amount of time, both for initializing Bioformats for import as well as the import itself for large images. The delay can cause the GUI to be unresponsive if not in a separate thread and may also be undesirable if the user does not wish to import this file.

Instead of importing the file automatically, switch tabs to the import panel if the file cannot be loaded. Initiate that import setup using the separate import thread to retain GUI responsiveness and show the user the files and settings that will be imported. To switch tabs programmatically, add a Trait to specify the tab to select based on an Enum class mapping control panel tabs to indices.