Validate shape and data type before allowing import
Check the output image shape and data bit and type before enabling the import button. The values should be considered valid when the fields have been auto-populated from extracted metadata but invalid when metadata is not available such as for RAW files. Change the import setup feedback based on whether this metadata has been supplied while dynamically changing the import button status based on the current entries of the corresponding fields. Also, remove the "data" from "data type" in the type default entry to reduce redundancy with the line's "data" label and avoid confusing this field with the entire data type. Add "order" to the byte order field so that the default entry shows what the field is for.