Reduce image adjustment slider label sizes
The image adjustment slider ranges have been based on the underlying data type. The number of digits for these data types can quickly enlarge labels to the point that the sliders themselves are narrow. Limit the label sizes by formatting the numbers to use scientific notation if beyond a string length limit.