Narrow image adjustment intensity range
Using the data type range to set the intensity ranges can lead to excessively large ranges for the given data, which makes fine adjustment by the user difficult. Use the pre-calculated near min/max values generated during import from the full image5d if available to limit the data range, continuing to fall back to the data range if these pre-calculated values are not available.

At times these near min/max values may not cover the range pre-selected values or the image range. Although the min/max values of the entire image cannot be loaded without needing to load the full image, the range for the given plane can be measured quickly. When synchronizing the sliders with the given image, expand the slider ranges to fit the pre-selected values and given plane's range if necessary.