Adjust overlapping intensities based on changing value
When decreasing a maximum value below the current minimum value in the image adjustment sliders, the minimum value takes on the maximum value, but the minimum value has been simply decreased to the maximum value when attempting to exceed it. Match min and max values based on the currently changing value so that the minimum value can push up the maximum value.