Move blob detection controls into a separate panel
Shift controls for blob detection into a new panel as these controls are distinct from the ROI, and the ROI panel has become overcrowded. Move the detection feedback text box to this panel and add a separate feedback box back to the ROI panel for its own feedback. Also, fix a typo in the import panel objective magnification label.