Fix intensity sliders when near max was not loaded
The default near maximum is a list of -1, causing the maximum intensity limit and current intensity negative values. Use the data type maximum if the near max value is negative, assuming that loaded near max values should be positive. Also, fix attempting to adjust the display settings for a non-existing editor, which was seen when the negative intensity limit triggered this adjustment.