Simplify overlaid images default parameter setup
Vmin/vmax and alpha parameters in the image overlay function have only been given default values when the parameter is None. Expand this default parameter setup for values of None for the first image, which has been treated as a special case anyway as the main image. This change obviates the need for the complicated inner function that had to support several differing use-cases. Also, clarify the comments on the structure of the channel parameter.