Remove `run_cli.py` wrapper script
Integrate non-GUI CLI functionality into the run script by assuming that the CLI will shut down the process before attempting to the run the GUI if the CLI performed a user-specified command-line driven task. Previously one loss of functionality from this approach was that windows were no longer integrated as tabs in a single window on macOS, but the integrated GUI precludes this requirement. This change also simplifies access through a single launch script. To avoid duplicating CLI argument setup, call it in the main GUI module only when it is running in the main scope.