Wrapper script to launch when `python` is not an available command
The `run.py` launch script assumes that `python` is available as a command when executing the script as a text file based on its shebang. On Linux such as Ubuntu 20.04, only `python3` may be available, meaning that this script cannot be launched through `./run.py` or a file browser. Also, platforms where `python` is only available through Conda may not be able to identify this command if the full Bash profile is not loaded, which may occur when launching the script through a file browser.

To support these use-cases, provide a Bash wrapper script that launches `run.py` with `python3` if this command is available. If not, the wrapper attempts to run the script through a `conda run` command in an interactive Bash shell. The interactive Bash shell makes `conda` accessible if it has been initialized or added to the path, and `conda run` makes the Conda `python` accessible for running the launch script.