Use an interactive Bash shell in the run script to access configuration
Some platforms such as Linux, the Bash configuration may be bypassed during non-interactive loading such as launching a Bash script from a file browser. To load this configuration, run commands in an interactive Bash session on non-Windows platforms in the run script. Continue to use the system shell on Windows since Bash is typically not available.