Find existing Conda installations more aggressively during Conda env setup
The Conda setup script attempts to install Miniconda if the `conda` command is not found. Expand checks for existing Conda installations by accessing the `CONDA_EXE` environment variable or `which conda` in interactive Bash sessions, since the configuration required for accessing Conda may not be available in non-interactive sessions such as launch from a file browser.