Generalize Bash script for alternate Pythons to include vanilla `python`
The Bash script wrapper for the Python run script has covered the Python binary cases that the run script would not handle by itself, where Python is only present as `python3` or from within Conda. To generalize this wrapper, include `python` as well in case users use this wrapper in place of calling the run script directly.