Replace copying run script with simple Bash wrapper
The Conda setup script has copied the appropriate run script based on whether the `python` command is present in the system. This detection may not work, however, if setup were run in a fully configured shell (ie with `.bashrc` fully loaded), where `python` could be seen, but the run script were launched non-interactively, where `python` might only be available through Conda. Copying the file also has the downside of not getting updates to the source file unless the user reruns the Conda setup.

As a simpler method that also remains updated, replace this copying with generating a simple Bash script pointing to the wrapper Bash run script. This change serves several functions:
- Generating a new file rather than having a static file avoids confusion for non-Bash platforms, such as Windows where the run script currently needs to be run directly
- The script itself is trivial
- The script points to another script, obviating the need for a soft link, which does not take advantage of the default Terminal file association in Mac, or a hard link, which may not follow updates if the source file is destroyed and recreated rather than directly edited