Workaround for ROI Editor title overlap
The ROI Editor figure title frequently overlaps the overview plots slightly since titles are not accounted for in the standard, non-constrained layout, and the constrained layout has a performance impact. To reduce the chance of this overlap, widen the margin at the top of the figure. Also, squeeze the title text into a single line. Change the background to the same as that of the Plot Editor label text to provide some contrast with the background and for consistency with the Plot Editor.