Show a preview of the ROI at a newly clicked position
The ROI has typically been positioned by offset sliders and text boxes for the size. Now that the user can adjust the offset by clicking on the overview plots in the ROI Editor, provide a preview of the ROI at the clicked position by showing a lighter, dotted box. This box is repositioned with each click.