Move up Conda environment setup in Dockerfile
Reorder the Dockerfile configuration so that the Conda environment setup command comes earlier, which allows this command to take advantage of the cache instead of triggering a rebuild with any other file changes. Split out the `environment.yml` copy command to precede this Conda setup so that any environment changes will trigger a rebuild of the Conda environment.