Fix indexing channels when only select channels are displayed
The Plot Editor has assumed that the given channel index matches the corresponding image in the saved image plots. If only a selected subset of channels is shown, however, the channel index is out of sync. To save the state of the currently displayed images' channel indices, build up the channel indices directly when showing an overview image, similar to the other image parameters. When getting a displayed image, translate the channel index to its position in the list of displayed indices. Return None if not found, such as when the a channel check box has been selected without redrawing the images.