Double-click in an ROI Editor overview plot to trigger a full redraw
Single clicking in an overview plot shows a preview of a new ROI at the clicked position. To ease redrawing the full ROI stack at that spot, provide an event handler for double-clicking in an overview plot. Add an attribute to take a redraw function that will redraw the editor for the given ROI for these events.