Change ROI Editor redraw shortcut to double right-click
In the double left-click shortcut introduced in commit de0ccffd, the first click of the double-click triggered an ROI reposition event before the editor was redrawn at that position. This behavior prevents the user from using the preview ROI for the exact placement of the new ROI since the ROI will be repositioned at the mouse position of the double-click, which is likely not to correspond exactly to the ROI preview position. There does not appear to be a way to ignore this first click in the Matplotlib API (see https://stackoverflow.com/questions/53533932/python-pyplot-handling-double-click-event-catches-also-the-first-click-event). Instead, change the shortcut to a double right-click since the ROI is not repositioned by right-clicks.