Change the ROI Editor overview plot navigation shortcuts
Based on feedback from users, users frequently wish to use the right/left arrows to shift planes in the overview plots since these arrow directions correspond to the layout of z-planes in the ROI plots. Add these arrows in addition to the up/down arrows used for overview plot navigation.

The right arrow + left-click combination while inside an ROI plot has not been intuitive, and the right arrow now conflicts with the the overview plot navigation. As a simpler shortcut, replace this combination shortcut with simply right-clicking on an ROI plot to jump to the corresponding plane in the overview plots.