Fix showing registered images for multiple suffixes for a registered image type
Overlaying multiple registered images of a given type, such as the registered atlas volume and sample image, sets the registered suffixes for the atlas type to a sequence. This sequence causes an exception when extracting the extension from the suffix. Handle both sequences and non-sequences, simply taking the first value for sequences.