Remove `image5d` from main registered image list
The `image5d` entry was supposed to be the default image, but this image is typically not a "registered" image such as the "exp.mhd" file. Selection of this value also does not have any effect. Remove this entry and instead assume that if no registered intensity image is selected, `image5d` will be shown. Change the label from "Main image" to "Intensity" to clarify that this image is not necessary the main image shown, but the registered intensity image.

Also, clear the saved register suffixes when the registered images are reloaded with no images selected. This change allows `image5d` to be restored.