Specify the `pyqt` Conda package explicitly
The `pyqt` package has been left unspecified in the Conda environment specification script since the package is a dependency of other dependencies. Recently an older version of PyQt5 (5.9) has been downloaded and installed for unclear reasons, whereas specifying the package in the spec causes the newer version (5.12) to be picked up instead. Since this package is required anyway, add it to the specification file.