Refresh viewer tabs upon selection after a new image has been loaded
The viewer tabs are redrawn when initially loaded or if the user explicitly triggers a redraw event. When opening a new image, the currently displayed viewer is redrawn, but the other viewers are not, which causes them to shown old contents when opened. While all viewers could be redrawn when loading a new image, this loading may unnecessarily impact performance, especially if the user does not end up opening those viewers. Instead, defer redraws by adding a stale flag to the viewers to trigger redraws when a viewer tab is opened after a new image has been loaded.

For simplicity, combine the viewer tab checks in the tab selection handler since any redraw will call the same function. Although the ROI Editor has been assumed to have been initialized since it is the initially selected tab, it is possible that the user will navigate away from this tab without a loaded image, such as when exploring the GUI. Check the ROI Editor's initialization, making the corresponding variable public so that it is accessible from within the handler.