`Blobs` class for storing blob-related attributes
Create a separate class for blobs to store blobs and related attributes such as blob matches or associated arrays with matching order to blobs. Refactor blob matches into instances of this class. Pass an instance of this class to the ROI Editor rather than the individual blob components. Refactor `segments` in the ROI Editor to `blobs` for clarity.