Use the full YAML profile path when checking the `profiles` folder
The YAML path had been truncated to the basename when checking the `profiles` folder, but getting the basename is not necessary to find a path in this folder since the full path will be checked later. To allow the user to specify a path in a sub-directory of the `profiles` folder without specifying the `profiles` folder itself, simply use the full path from the start. Also, fix the error message to use this original path when informing the user of the next path to be checked.