Avoid unintended arguments to blob detection handler
The blob detection handler gets triggered with TraitsUI arguments, which causes unintended arguments to the `segs` parameter. Create a separate function for the blob handler, which serves as a wrapper with no arguments to the blob detection function. This change avoids the need for the private function checking for specific values of the `segs` argument.