Provide an option to allow the found parent to be at same level as its child
Provide a workaround for discrepancies between the listed parent in the ontology reference file and the parent derived based on a target level numerically below the child. Add a parameter allowing a parent to be selected so long as it is either at or below the numerical level of the child. The parameter is turned off by default, but the only current implementation turns it on for consistency with the ABA reference file. Also, change the docs convention to describe a level as "below" to mean numerically lower for consistency with the level values.