Add a command-line parameter for loading data files
The blob load parameter has been incorporated into the task processing (`--proc`) parameter but is not a processing task and may be needed with other processing tasks. Other data files may also need to be loaded, but the parameter currently refers only to blobs. To allow greater flexibility when specifying data files to load, add a `--load` command-line parameter.

Extend `ars_to_dict` to take a default value so that the arguments can be given simply by name, without concern for order or value assignment. This default value can be useful for arguments that simply assign True to any given keyword, for example, such as whether or not to load the given data sources.