Provide a file dialog to choose path and file format before saving figure
The button to save a figure in the main GUI has saved to a default path using an extension/file format based on the `--savefig` command-line argument or defaulting to PNG. To provide more flexibility in save location and file format, open a file browser to select a file before saving the figure. Continue to provide a default path based on the currently selected editor, prepopulating the file dialog with this path. By allowing the user to change the path, the file format can be changed without using the command-line argument.

Additionally:
- Fix exception when attempting to save a 3D image when no image has been loaded
- Provide feedback when there is no figure to save
- Split default path creation into separate function from saving the figure