Fix error when Javabridge is installed but Java is not
If the Javabridge module are installed but Java is not installed, Javabridge will issue an `javabridge.jutil.JVMNotFoundError` rather than an `ImportError`, leading to a separate exception. Catch this exception to allow the software to be run Javabridge but not Java. Catch the exception as a `ValueError` since this error cannot be imported without triggering the same exception.