Docs on workaround for long Java load times
The older version of Java installed in the Conda setup pathway takes longer to initialize than newer versions do but provides backward compatibility for other software such as ImageJ used in the stitching pipeline. Provide instructions on a workaround to replace this Java with a newer version. Also, add sub-headers within the section on long load times.