Fix mapping labels with discrete colormap to binary colors
Label binarization maps foreground to NaN to use a given color for all NaN values, but foreground and background are not obtained by simple 0 vs non-0 distinction after label conversion for discrete colormaps. Shift this conversion until after changing foreground pixels to NaN. Note that the NaN color is still not applied since the converted label is not NaN, but the length of the label indices.