Include filter size in smoothing metric data frames
Add a parameter to the label smoothing metric function to take the filter size and include it as a column in the data frame of smoothing metrics for each label. This addition distinguishes the metrics when multiple data frames for smoothing at different filter sizes are combined. Incorporate this column in the aggregated smoothing metrics of any given data frame as well, assuming that the filter size will be the same for all labels and thus the average filter size is the same.