GUI option check boxes for region selection
Add check boxes for region label selections. Choosing both sides of a label has required using the "+/-" notation, but now the user can select the corresponding check box. The notation is still supported for backward compatibility and also to allow selecting both sides only for specific labels, in which case the check box should be left unselected.

Another option added is whether to include children of the labels. The default of including children remains the same, with the box selected by default.