Require at least some erosion before stopping during label erosion
Labels that do not undergo erosion have stopped undergoing erosion with smaller filter size and are reported to have been filtered with the given size. This behavior may be seen with particularly small labels since the `binary_erosion` function from Scikit-image treats outside (border) pixels as True, meaning that labels of a single pixel for example will not be eroded. By reporting this filter size, the label is not excluded from the downstream watershed step.

To more correctly report that the label has not been eroded, require the label to decrease in volume before stopping the erosion. The subsequent filter trials will likely also not decrease the volume since filter sizes decrease, but simply continue these steps for simplicity until reaching the minimum filter handling step.