Parameter to support 3- vs 5-based images when preparing ROIs
While most image arrays in MagellanMapper assume a time dimension, registered images (eg labels) have been assumed to omit this dimension. To specify whether this axis is present or not when extracting an ROI or sub-image from a larger image, add a parameter to specify the number of dimensions on which the image is based.