Implement label translation when comparing values
Add `--atlas_labels translate_labels` as a label translations command-line argument to specify CSV files with label identity translations. To allow translation for any given image, take a list of paths corresponding to the list of labels images. Provide a warning if a CSV file does not exist while allowing an empty string to skip translation for a given image.