Fix retaining references to unused Atlas Editors
Launching an Atlas Editor in Visualization stores the Atlas Editor instance in an array to access multiple Atlas Editors that are open simultaneously. Currently, multiple Atlas Editors are not supported now that the editor has been integrated into the unified interface, although this option may be re-introduced later. Redrawing the Atlas Editor creates a new editor object to obtain a clean refresh, and thus the old object should no longer be referenced to allows its resources to be garbage collected.