Provide an option to treat the ROI offset as the center of the ROI
Add a boolean check box to "center" the ROI, meaning that the ROI offset will be treated as the center of the ROI rather than its upper left corner. This change allows the user to specify ROI coordinates based on their target for the middle of the ROI, which may be more intuitive for some users. When this box is checked, retrieving the ROI offset values stored in the user controls will convert them from "center" values to "upper left hand" (ie true offset) values except for the Atlas Editor since it already centers its planes at the given offset.