Refactor ROI and Atlas Editors to use a common Plot Editors attribute
Use the same attribute to store Plot Editors from ROI and Atlas Editors in the Mixin from which they both inherit. This change allows the Mixim image display functions to be used directly instead of requiring wrapper functions. The attribute is an `OrderedDictionary` to encapsulate the functionality in the Atlas Editor, where planar orientations map to the Plot Editors, and the ROI Editor now maps zoom levels to Plot Editors, although the zoom levels are not currently used.