Option to zoom into the ROI in the Atlas Editor
Provide a check box to zoom the Atlas Editor into the ROI. When the "Center" ROI option is unselected, the crosshairs are considered to be the ROI offset, zooming to place the crosshairs at the image plane borders. If "Center" is selected, the crosshairs are treated as the center of the ROI and thus are centered in each plane. Also, make the y-limit order when zooming into a Plot Editor configurable since the need to reverse this order depends on the axis.