Create a blank labels image if the given labels image path is not found
Provide a way to generate a blank image to save to a new path. When the user supplies a registered image suffix for the annotation image that does not exist, a blank image will be generated using the same shape as that of the x,y,z axes of the main image. Saving the labels image in the Atlas Editor will save it to the new path. Also, the labels reference file is now loaded even if the labels image file is not loaded so that any labels added to the blank image will show the corresponding label descriptions.