Support R plotting without an available screen device
R plots have been unable to output to file when through through Rscript. To support running R stats both through Rscript and directly in an R session, where the screen device is available, first attempt to set up a PDF device prior to plotting. If a screen is not available to set up this device, open a PDF device to output to file directly.

Also, add a function to reset the device, though currently this reset appears to interfere with graphics output even if called before the graphics have been set up or after apparent completion.