Ensure that atlas and labels have same planes when not mirroring atlas
Turning off atlas mirroring while still mirroring labels could result in atlas and labels images with different numbers of planes since mirroring often does not take place at midline. To ensure that the two images have the same number of planes, pad the image with the fewer number of planes to enlarge it to the size of the other image.