Group `export_stack` imports in `cli`
The `export_stack` module has been imported only just prior to processing task to avoid loading a graphical session during non-graphical tasks. Now that a graphical session is no longer loaded when this module is imported, group its import with the rest of the imports in the `cli` module, following the convention of specifying imports at the top of the file. Also, group imports from the same sub-package on the same line and alphabetize sub-packages.