Option to weight edge extension by lateral plane position
Provide an option to reduce weighting of medial planes for erosion filters so that erosion becomes progressively stronger when while extending farther laterally. Add the atlas profile setting, "wt_lat", within the edge extension dictionary to configure this setting as a fraction by which to apply this weighting. Determine the position within the lateral extension based on the anatomical edge map. Also, clarify more of the edge extension settings in the atlas profile comments.