Support custom prefixes and suffixes when making an output path
The function to create an output path incorporates the config prefix/suffix attributes, but occasionally a custom set of values may be desired. Add parameters to take these values, falling back to the config values if the parameters are not given. Also, simplify the path building by simply returning if a prefix is given. Apply these new parameters to line and scatter plot functions.