Avoid attempting to save a figure to an unsupported format
If no save extension is given, the output path is sometimes the same as the input path, such as a CSV file. Attempting to save the figure to this file causes an error. Prior to saving, the original file is also backed up, changing its location and making it appear missing. To avoid this error and backup, check the supported file formats for the figure's backend before attempting to save the file. If the file extension is not supported, print a warning and return before backing up the file.