Avoid repeated file saving during 2D plot tasks
2D plot tasks have typically saved files during the individual task, but the file save does not include the post-processing step, which creates an additional file. To avoid repeated saves, turn on saving in the individual tasks, deferring saving to the post-processing task now that this task should run anytime a plot is successfully generated.

This post-processing step has only saved files if both an output path and save extension are specified. Out paths were previously only specified for a subset of tasks, but a generic path is now generated if none exists. Continue to save only if a save extension is given, which allows the user to specify not to save by setting `savefig` to None but defaults to saving now that `savefig` defaults to PNG.