Move R stats regions to ignore parameter to environment configuration
The ignore regions constant has been applied based on the input stats file, but now stats files can be specified through a command-line parameter and may have a different path or filename. Since the need to ignore regions is dependent on the analysis type rather than the input file, move this parameter to the environment configuration.