Fix jitter plots to generate plots in both interactive and non-interactive environments
Interactive R environments provide a screen device on which to plot without explicitly opening a device and to save afterward through `dev.print`. Non-interactive environments such as R instead require a device such as `pdf` to be opened explicitly, which in turn prevents output to an interactive environment. To accommodate both environments, check for interactivity and change the output device handling accordingly.

- In interactive environments, open the device before setting graphical parameters so that they are adopted in the current device and plot
- Use a `tryCatchLog` for the plot to provide the full stacktrace in case of errors
- Fix display in interactive environments by not opening a non-screen device and by resetting the device after each plot
- Fix saving in interactive environments by saving after the plot has been fully generated