Return merged data frame when combining CSV files
The function to merge CSV files has been used only as a command-line task for outputting to file. Extend it for use as an API to return the merged data frame object. Make the output path parameter optional since the data frame may only be an intermediate object used for further processing. Also, update additional docstrings in the data frame module.