Parameter for sequence function to allow wrapping None in a sequence
The `to_seq` function has not wrapped a value of None in a sequence since None is often used in place of a sequence to indicate that the value should not be used or a default value should be used instead. Sometimes even None needs to be wrapped in a sequence, however, and might be expected for a `to_seq` function. Provide this option as a parameter, continuing the original behavior by default.