Workaround for incorrect color display for some corresponding labels
Positive and negative labels displayed with a discrete colormap for symmetric colors should show the same color. These colors are occasionally mismatched for unclear reasons, likely related to rounding or resampling issues in Matplotlib image display. The mapped colormap value is correct, but the displayed image is not. This issue only occurs with certain sets of values, potentially pointing to an issue with the image resampling range performed during Matplotlib image display. Indeed, one workaround is to change the data type during Matplotlib resampling from a `float32` to a `float64` type, but this change requires an external library change and increases memory usage.

As a simple alternative, take the absolute value of the image when converting image label values to their corresponding indices if the colormap is set to display symmetric colors. Make the symmetric colors flag an attribute to track this setting. The resulting converted image maps only to the positive values for consistent display across both positive and negative labels. Since the pixel display marker operates on the original image, the displayed label ID continues to show the correct signage.

(cherry picked from commit 188641de8cb4baea425b8fe306320bdc847990f2)