Fix reading size and resolution metadata when x, y, or z are missing
Once a x, y, or z value for size and resolution metadata is missing during multi-page file import, no further values for the parameter are read. To continue reading this information, provide a default value if the key is not found. Remove the try-catch block since the presence of the key is checked directly.