Add default parameter values for database experiment functions
The experiment selection and insertion functions can accept None values for experiment names and dates, and these parameters can be ignored under certain use cases. Add default values to avoid the need to set them when they are unused and to clarify that they accept None. Update the docstrings to clarify usage in the default cases.