Change the shape parameter used for whole image blob match ROIs
Processing blob matches on the whole image inserted an ROI whose dimensions covered the full image, but retrieving this ROI requires knowing the shape of this image, which may not be available in downstream analyses. To remove this requirement, use a shape of `(0, 0, 0)` instead. This change also has the benefit of immediately indicating that the ROI is for the whole image.

Since the colocalizer blob match insertion and selection functions are used only for whole image processing, remove the offset/shape parameter from the insertion function, instead using these zero values. For the selection function, use the offset/shape only for finding a subset of blob matches within an ROI while using the zero values to select the ROI. Also, make the selection offset/shape optional, where the default value simply takes all blob matches in the whole image ROI. This change shifts their order in function signature, and None is returned immediately if no blobs with matches are found.

ROI selection also requires the image name for the experiment name, but the image may not have been loaded if it was not available. Add a parameter to specify the name of the experiment in case it cannot be determined by the loaded image for ROI selection.