Support loading blobs from a user-supplied path
Change the command-line parser for `--load_data` to use a boolean as the default sub-argument value so that strings can be detected as user-supplied paths. Also, change the delimiter to a bar to allow commas in paths. Check for these user-supplied blobs archive paths when converting an image-based path to a blobs path, giving any user-supplied path precedence. This change allows users to supply a blobs archive path in the format, `--load_data blobs=<path-to-blobs-archive.npy>`.