Avoid unnecessary 3D viewer refreshes when displaying blobs
The 3D viewer may not need to undergo a full refresh when displaying blobs, unlike the current ROI Editor requirement. Only refresh the 3D viewer if it is shown and stale. While the viewer could be refreshed even if not shown, currently do not refresh it as this viewer can take a considerable amount of time and may not be necessary. Instead, defer to refresh when the user explicitly requests it or if shown for the first time.