Make 3D blobs pickable
Add a listener for mouse pick events when showing 3D blobs. Respond to blob pick events by changing the ROI outline to surround the picked blob instead. Reset the outline to the full ROI when clicking on a non-blob.

The mouse picker is inspired by the Mayavi ball pick example:
https://docs.enthought.com/mayavi/mayavi/auto/example_select_red_balls.html

with further explanations/fix from here:
https://stackoverflow.com/a/54806091/1911852