Pick the closest 3D blob when a blob is not picked
A 3D blob may not always be selected when attempting to click on one since blobs may be small relative to the ROI. To increase the chances of selecting a desired blob, check for the closest blob to a given click as inspired by https://stackoverflow.com/a/62556622/1911852. If the blob falls within a certain tolerated distance, select this blob. Otherwise, continue to fall back to changing the outline to the full ROI.