Option to sync Atlas Editor with ROI offset
Provide a check box in the Atlas Editor options to synchronize the editor with the ROI offset controls. Selecting this option causes the Atlas Editor to move positions as the ROI offset selectors are shifted, rather than waiting until the Refresh button is pressed. This movement serves a similar role as the sliders built into the Atlas Editor itself but keeps the ROI tab selectors synchronized with the editor. This option also allows changes to the ROI offset elsewhere such as 3D blob selection to update the Atlas Editor position. Defer updates to the editor when it is not currently viewed to avoid unnecessary image loading. This option is set on by default, but provide the option to turn if off to avoid loading images with every ROI offset adjustment, which may be preferred for large images where large image planes may take awhile to load.