Refactor 3D viewer options into enumerations
Change from list to Enum to improve readability of the given options. Rename the side panels and circles for brevity and clarity and move them to the end of the list as they are likely used less frequently. Also, reduce the number of columns in the Atlas Editor options check list, basing this value on the number of enumerations as with the 3D viewer options.