Provide an option to clear the scene before drawing new 3D object
Drawing 3D surfaces or point-representations of objects has cleared the scene beforehand. Leaving the scene intact would allow objects to be layered on top of one another, such as showing multiple separate atlas structures. Provide a 3D viewer option to toggle whether to clear the scene beforehand. Default to turning this option off to avoid accidental erasing. Also, prevent the scene from being cleared anytime the "raw" option is unchecked and the 3D scene is refreshed, instead only doing so if the "clear" button is checked to making this clearing explicit. Move the clearing call from the 3D surface/point functions to a separate clear function.