Fix ROI sync when using a centered ROI offset
Shifting the ROI in the Atlas Editor with both the ROI sync and offset "center" options selected leads to an infinite recursion because the new offset is converted to a centered offset. Fix by not converting the offset when syncing the ROI for the Atlas Editor. The Atlas Editor should receive the same ROI offset that it already has and will not recursively update the ROI.