Further prevent recursive ROI update handling
Programmatically updating the ROI offset can cause the ROI offset callback to recursively update itself when syncing with the Atlas Editor. Since the ROI offset callback is meant to handle user-directed changes to the offset slider controls, the callback can be largely ignored if the offset is changed programmatically instead. Add a flag to toggle whether to ignore ROI offset changes and ignore ROI sync when the offset is set programmatically, typically by the Atlas Editor itself.