Limit Atlas Editor options callback to changed settings
The callback for the Traits List objects responds to any changes in the list, meaning that the rest of the settings are unchanged each time the callback is triggered. The handler has been implemented so that resetting unchanged settings has minimal impact, but ignoring unchanged settings would further reduce any current or potential side-effect. Remove the show labels private function for simplification.