Integrate the ROI zoom control into the Atlas Editor options list
Rather than treating the zoom control as a separate option, integrate it with the rest of the Atlas Editor options list of check boxes. The option is specific to the Atlas Editor, and several other options also have immediate responses as the zoom option does. The integration also makes the option layout more coherent and consistent. Also, add the crosshairs description to the options tooltip.